Refactor: 공지 리스트 section을 별도의 컴포넌트로 분리
