+++ src/widgets/post/sections/NoticeList/index.ts
@@ -1,0 +1,1 @@
+export * from './NoticeList';
